Extract result serialization to file
