+++ Duplicati/Library/Modules/Builtin/SendHttpMessage.cs
@@ -321,7 +321,7 @@ namespace Duplicati.Library.Modules.Builtin {
             if (input.IndexOf("%RESULT%", StringComparison.OrdinalIgnoreCase) >= 0)
                 using (TempFile tf = new TempFile())
                 {
-                    RunScript.SerializeResult(tf, result);
+                    ResultSerializer.SerializeResult(tf, result);
                     input = Regex.Replace(input, "\\%RESULT\\%", System.IO.File.ReadAllText(tf), RegexOptions.IgnoreCase | RegexOptions.CultureInvariant);
                 }
